Fix installation requirements for development
If this package has optional dependencies (called "extras"), the proper
way to specify them is not in the requiremets.txt file. Any extras
specified here are on the required packages, not on this package.
Instead of going all the way to a proper build setup, take an
intermediate step, and isolate extra dependencies in a separate
requirements-dev.txt for developers.