Fix error with data not found by PyInstaller
This is a very hacky fix. The problem is that when running the exe file created by the PyInstaller, it is unable to find the factoriocalc\game-info-normal.json file. I believe PyInstaller doesn't insert it into the exe file, so the fix is manually including it as data in the bundle-python.sh. This is though not system specific for now, so only works on my laptop.