Switch server to Flask
This will reduce debugging time, since a new exe doesn't have to be built each time. This also reduces complexity. Importantly this helped fix a bug with context vars not being found. The error was that the server was threaded (note currently threaded = false). Apparently a threaded server can have problems having the data in the right location (on the right thread?).