Fix problem with path overlapping end inserter
That is, a path could be found, where part of the path was on the same squares, as the inserter, that inserts on to the path. This would result in a "already reserved" error, when attempting to place the inserter after the path of belts had been placed.
The problem was that there was no check if the path was valid before returning it. Now it tries to backtrace and checks if anything in the end is on an illegal square.